Make tpd an s3 object and add alpha_plot
Adding the s3 object `tpd` allows for automatic checking of the (assumed or estimated) alpha and throws a warning if a user attempts to use, e.g., innovations() on a `tpd` object without alpha = 2. In addition, adding the alpha_plot() function gives users a simple way to quickly check whether their data can be reasonably assumed to share an alpha.