Change alpha estimation in tpd()
We had used evd::fpot() to estimate a shared alpha which assumes a shared alpha and a shared "scale". This is too strong of an assumption. Instead, we now use the average of the Hill estimators for each variable using a fixed number of order statistics `k`. This assumes that a shared `k` is sensible and that the alpha is the same.